Added Images and Progress Bar

+++ app/src/main/java/com/rahul/weatherapp/model/WeatherResponse.kt
@@ -17,7 +17,7 @@ data class List(
     @SerializedName("clouds") var clouds: Clouds? = Clouds(),
     @SerializedName("wind") var wind: Wind? = Wind(),
     @SerializedName("visibility") var visibility: Int? = null,
-    @SerializedName("pop") var pop: Int? = null,
+    @SerializedName("pop") var pop: Double? = null,
     @SerializedName("sys") var sys: Sys? = Sys(),
     @SerializedName("dt_txt") var dtTxt: String? = null
 )
